backend/requirements.txt
# ArchitectAI API — pinned for production
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Agent & LLM
langgraph>=0.2.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langchain-openai>=0.2.0,<1.0.0

# HTTP client (GitHub API)
httpx>=0.27.0,<1.0.0
# Rate limiting
slowapi>=0.1.9,<1.0.0
# Request tracing
asgi-correlation-id>=4.3.0,<5.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
aiosqlite
asyncpg

# Auth
python-jose[cryptography]
bcrypt>=4.0.0,<5.0.0

# Payment Gateway
razorpay>=1.4.1,<2.0.0

# Image processing (for watermarks)
Pillow>=10.0.0,<11.0.0
